+++ packages/core/src/mdx-plugins/remark-code-tab.ts
@@ -1,9 +1,11 @@
 import type { Processor, Transformer } from 'unified';
-import type { Code, Root } from 'mdast';
+import type { BlockContent, Code, DefinitionContent, Root } from 'mdast';
 import { visit } from 'unist-util-visit';
 import type { MdxJsxFlowElement } from 'mdast-util-mdx-jsx';
-
-const TabRegex = /tab="(.+?)"/;
+import {
+  generateCodeBlockTabs,
+  parseCodeBlockAttributes,
+} from '@/mdx-plugins/codeblock-utils';
 
 type TabType = keyof typeof Types;
 export interface RemarkCodeTabOptions {
@@ -15,6 +17,12 @@ export interface RemarkCodeTabOptions {
    * @defaultValue false
    */
   parseMdx?: boolean;
+}
+
+declare module 'mdast' {
+  export interface CodeData {
+    tab?: string;
+  }
 }
 
 const Tabs = {
@@ -143,64 +151,28 @@ const CodeBlockTabs = {
     withParent = true,
   ): MdxJsxFlowElement {
     const names = processTabValue(nodes);
-    const children: MdxJsxFlowElement[] = [
-      {
-        type: 'mdxJsxFlowElement',
-        name: 'CodeBlockTabsList',
-        attributes: [],
-        children: names.map((name) => {
-          return {
-            type: 'mdxJsxFlowElement',
-            name: 'CodeBlockTabsTrigger',
-            attributes: [
-              {
-                type: 'mdxJsxAttribute',
-                name: 'value',
+    const node = generateCodeBlockTabs({
+      defaultValue: names[0],
+      triggers: names.map((name) => ({
+        value: name,
+        children: [
+          withMdx
+            ? // eslint-disable-next-line @typescript-eslint/no-explicit-any -- needed
+              (mdxToAst(processor, name) as any)
+            : {
+                type: 'text',
                 value: name,
               },
-            ],
-            children: [
-              withMdx
-                ? // eslint-disable-next-line @typescript-eslint/no-explicit-any -- needed
-                  (mdxToAst(processor, name) as any)
-                : {
-                    type: 'text',
-                    value: name,
-                  },
-            ],
-          };
-        }),
-      },
-      ...nodes.map((node, i) => {
-        return {
-          type: 'mdxJsxFlowElement',
-          name: 'CodeBlockTab',
-          attributes: [
-            {
-              type: 'mdxJsxAttribute',
-              name: 'value',
-              value: names[i],
-            },
-          ],
-          children: [node],
-        } as MdxJsxFlowElement;
-      }),
-    ];
+        ],
+      })),
+      tabs: nodes.map((node, i) => ({
+        value: names[i],
+        children: [node],
+      })),
+    });
 
-    if (!withParent) return createFragment(children);
-
-    return {
-      type: 'mdxJsxFlowElement',
-      name: 'CodeBlockTabs',
-      attributes: [
-        {
-          type: 'mdxJsxAttribute',
-          name: 'defaultValue',
-          value: names[0],
-        },
-      ],
-      children,
-    };
+    if (!withParent) return createFragment(node.children);
+    return node;
   },
 };
 
@@ -216,10 +188,10 @@ export function remarkCodeTab(
   const { parseMdx = false, Tabs = 'CodeBlockTabs' } = options;
 
   return (tree) => {
+    const ignored = new WeakSet();
+
     visit(tree, (node) => {
-      if (!('children' in node)) return;
-      let start = -1;
-      let i = 0;
+      if (!('children' in node) || ignored.has(node)) return 'skip';
       let localTabs: TabType = Tabs;
       let localParseMdx = parseMdx;
       let withParent = true;
@@ -233,7 +205,7 @@ export function remarkCodeTab(
         localTabs = node.name as TabType;
 
         // for `Tabs` in simple mode, it doesn't support MDX tab names
-        if (node.name === 'Tabs') {
+        if (node.name === 'Tabs' && localParseMdx) {
           localParseMdx = node.attributes.every(
             (attribute) =>
               attribute.type !== 'mdxJsxAttribute' ||
@@ -242,50 +214,49 @@ export function remarkCodeTab(
         }
       }
 
-      while (i < node.children.length) {
-        const child = node.children[i];
-        const isSwitcher =
-          child.type === 'code' && child.meta && child.meta.match(TabRegex);
+      let start = -1;
+      let end = 0;
+      const close = () => {
+        if (start === -1 || start === end) return;
+        const replacement = Types[localTabs].convert(
+          this,
+          node.children.slice(start, end) as Code[],
+          localParseMdx,
+          withParent,
+        );
 
-        if (isSwitcher && start === -1) {
-          start = i;
+        ignored.add(replacement);
+        node.children.splice(start, end - start, replacement);
+        end = start;
+        start = -1;
+      };
+
+      for (; end < node.children.length; end++) {
+        const child = node.children[end];
+        if (child.type !== 'code' || !child.meta) {
+          close();
+          continue;
         }
 
-        // if switcher code blocks terminated, convert them to tabs
-        const isLast = i === node.children.length - 1;
-        if (start !== -1 && (isLast || !isSwitcher)) {
-          const end = isSwitcher ? i + 1 : i;
-          const targets = node.children.slice(start, end) as Code[];
-          const replacement = Types[localTabs].convert(
-            this,
-            targets,
-            localParseMdx,
-            withParent,
-          );
-
-          node.children.splice(start, end - start, replacement);
-          if (isLast) break;
-          i = start + 1;
-          start = -1;
-        } else {
-          i++;
+        const meta = parseCodeBlockAttributes(child.meta, ['tab']);
+        if (!meta.attributes.tab) {
+          close();
+          continue;
         }
+
+        if (start === -1) start = end;
+        child.meta = meta.rest;
+        child.data ??= {};
+        child.data.tab = meta.attributes.tab;
       }
+
+      close();
     });
   };
 }
 
 function processTabValue(nodes: Code[]) {
-  return nodes.map((node, i) => {
-    let title = `Tab ${i + 1}`;
-
-    node.meta = node.meta?.replace(TabRegex, (_, value) => {
-      title = value;
-      return '';
-    });
-
-    return title;
-  });
+  return nodes.map((node, i) => node.data?.tab ?? `Tab ${i + 1}`);
 }
 
 /**
@@ -305,7 +276,9 @@ function mdxToAst(processor: Processor, name: string) {
   return node;
 }
 
-function createFragment(children: MdxJsxFlowElement[]): MdxJsxFlowElement {
+function createFragment(
+  children: (BlockContent | DefinitionContent)[],
+): MdxJsxFlowElement {
   return {
     type: 'mdxJsxFlowElement',
     name: null,